pyproject.toml
[project]
name = "query"
version = "0.2.0"
description = "Open Food Facts: The Open Food Facts Data Warehouse"
authors = [
    {name = "Open Food Facts Team"}
]
license = {text = "GNU Affero General Public License v3"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "motor (>=3.7.0,<4.0.0)",
    "fastapi (>=0.115.9,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "redis[hiredis] (>=5.2.1,<6.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
fastapi-cli = "^0.0.7"
testcontainers = {extras = ["postgres", "redis"], version = "^4.9.1"}
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
black = "^25.1.0"
autoflake = "^2.3.1"
isort = "^6.0.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.coverage.run]
omit = [
    # omit test files
    "*_test.py",
    "test_*.py",
    "query/conftest.py"
]

